requirements.txt
attrs==19.3.0
Automat==20.2.0
beautifulsoup4==4.9.1
cffi==1.14.1
constantly==15.1.0
cryptography==3.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
Twisted==20.3.0
zope.interface==5.1.0
